Create contentComponent
